Created one page for config and another for running flows.
UI to run flows is shown after successfully testing a connection to the terminal.
